refactor: split external services into dedicated packages
- Reorganized the external services module into three distinct packages: `external-services-firebase`, `external-services-paypal`, and `external-services-podcast-index`.
- Updated all relevant import statements and build scripts to reflect the new package structure.
- Adjusted documentation to outline the new module dependency order and build requirements.
- Removed the old `external-services` package and its associated files to streamline the codebase. #41
