+++ .llm/history/active/external-services-split/external-services-split-part-01.md
@@ -1,0 +1,43 @@
+# Feature: external-services-split (Part 1)
+
+> **Note**: This LLM history file is optional. If you're not using LLM assistance for development, you can delete this file and the containing directory. The history tracking system helps document LLM-assisted decisions but is not required for contributing.
+>
+> **10-Session Limit**: Each part file is limited to 10 sessions. When adding Session 11, create `external-services-split-part-02.md`.
+
+## Metadata
+
+- Started: 2026-02-01
+- Completed: In Progress
+- Author: Mitch Downey
+- LLM(s): Cursor, Claude, etc.
+- GitHub Issues: https://github.com/podverse/podverse/issues/41
+- Branch: chore/external-services-split
+- Origin: git@github.com:podverse/podverse.git
+- Is Fork: no
+
+## Context
+
+[What problem does this solve? What's the goal?]
+
+## Sessions
+
+### Session 1 - 2026-02-01
+
+#### Prompt (Developer)
+
+[First prompt will go here]
+
+#### Key Decisions
+
+- [Decision and rationale]
+
+#### Files Changed
+
+- [List of files]
+
+---
+
+## Related Resources
+
+- [Link to PR]
+- [Link to related issues]
