+++ apps/api/package.json
@@ -9,9 +9,9 @@
     "build": "npm run lint && tsc && tsc-alias",
     "build:prod": "tsc -p tsconfig.prod.json && tsc-alias -p tsconfig.prod.json",
     "dev": "tsc && tsc-alias && node ./dist/index.js",
-    "dev:watch": "tsc --build --watch --preserveWatchOutput & tsc-alias --watch & nodemon --delay 1000ms --watch dist --watch ../../packages/helpers/dist --watch ../../packages/external-services/dist --watch ../../packages/orm/dist --watch ../../packages/notifications/dist --watch ../../packages/parser/dist --watch ../../packages/mq/dist dist/index.js",
+    "dev:watch": "tsc --build --watch --preserveWatchOutput & tsc-alias --watch & nodemon --delay 1000ms --watch dist --watch ../../packages/helpers/dist --watch ../../packages/external-services-firebase/dist --watch ../../packages/external-services-paypal/dist --watch ../../packages/external-services-podcast-index/dist --watch ../../packages/orm/dist --watch ../../packages/notifications/dist --watch ../../packages/parser/dist --watch ../../packages/mq/dist dist/index.js",
     "dev:inspect": "tsc && node --inspect ./dist/index.js",
-    "dev:inspect:watch": "tsc --build --watch --preserveWatchOutput & tsc-alias --watch & nodemon --inspect --delay 1000ms --watch dist --watch ../../packages/helpers/dist --watch ../../packages/external-services/dist --watch ../../packages/orm/dist --watch ../../packages/notifications/dist --watch ../../packages/parser/dist --watch ../../packages/mq/dist dist/index.js",
+    "dev:inspect:watch": "tsc --build --watch --preserveWatchOutput & tsc-alias --watch & nodemon --inspect --delay 1000ms --watch dist --watch ../../packages/helpers/dist --watch ../../packages/external-services-firebase/dist --watch ../../packages/external-services-paypal/dist --watch ../../packages/external-services-podcast-index/dist --watch ../../packages/orm/dist --watch ../../packages/notifications/dist --watch ../../packages/parser/dist --watch ../../packages/mq/dist dist/index.js",
     "lint": "eslint ./src --max-warnings 0",
     "lint:fix": "eslint ./src --fix",
     "start": "node ./dist/index.js",
@@ -20,7 +20,9 @@
   },
   "license": "AGPL-3.0",
   "dependencies": {
-    "@podverse/external-services": "*",
+    "@podverse/external-services-firebase": "*",
+    "@podverse/external-services-paypal": "*",
+    "@podverse/external-services-podcast-index": "*",
     "@podverse/helpers": "*",
     "@podverse/mq": "*",
     "@podverse/notifications": "*",
