+++ apps/management-api/Dockerfile
@@ -39,8 +39,6 @@ COPY --from=builder /opt/packages/helpers-backend/package.json ./packages/helper
 COPY --from=builder /opt/packages/helpers-backend/dist/ ./packages/helpers-backend/dist/
 COPY --from=builder /opt/packages/helpers-config/package.json ./packages/helpers-config/
 COPY --from=builder /opt/packages/helpers-config/dist/ ./packages/helpers-config/dist/
-COPY --from=builder /opt/packages/external-services/package.json ./packages/external-services/
-COPY --from=builder /opt/packages/external-services/dist/ ./packages/external-services/dist/
 COPY --from=builder /opt/packages/orm/package.json ./packages/orm/
 COPY --from=builder /opt/packages/orm/dist/ ./packages/orm/dist/
 COPY --from=builder /opt/packages/notifications/package.json ./packages/notifications/
