+++ apps/workers/Dockerfile
@@ -39,8 +39,10 @@ COPY --from=builder /opt/packages/helpers-backend/package.json ./packages/helper
 COPY --from=builder /opt/packages/helpers-backend/dist/ ./packages/helpers-backend/dist/
 COPY --from=builder /opt/packages/helpers-config/package.json ./packages/helpers-config/
 COPY --from=builder /opt/packages/helpers-config/dist/ ./packages/helpers-config/dist/
-COPY --from=builder /opt/packages/external-services/package.json ./packages/external-services/
-COPY --from=builder /opt/packages/external-services/dist/ ./packages/external-services/dist/
+COPY --from=builder /opt/packages/external-services-firebase/package.json ./packages/external-services-firebase/
+COPY --from=builder /opt/packages/external-services-firebase/dist/ ./packages/external-services-firebase/dist/
+COPY --from=builder /opt/packages/external-services-podcast-index/package.json ./packages/external-services-podcast-index/
+COPY --from=builder /opt/packages/external-services-podcast-index/dist/ ./packages/external-services-podcast-index/dist/
 COPY --from=builder /opt/packages/orm/package.json ./packages/orm/
 COPY --from=builder /opt/packages/orm/dist/ ./packages/orm/dist/
 COPY --from=builder /opt/packages/notifications/package.json ./packages/notifications/
