+++ apps/workers/package.json
@@ -13,26 +13,27 @@
     "clean": "rm -rf dist",
     "type-check": "tsc --noEmit",
     "start": "node ./dist/index.js",
-    "archive_all": "ts-node -r tsconfig-paths/register ./src/index.ts archiveAll",
-    "mq_rss_add": "ts-node -r tsconfig-paths/register ./src/index.ts mqRSSAdd",
-    "mq_rss_add_all": "ts-node -r tsconfig-paths/register ./src/index.ts mqRSSAddAll",
-    "mq_rss_add_recently_updated_feeds_from_podcast_index": "ts-node -r tsconfig-paths/register ./src/index.ts mqRSSAddRecentlyUpdatedFeedsFromPodcastIndex",
-    "mq_rss_run_live_item_listener": "ts-node -r tsconfig-paths/register ./src/index.ts mqRSSRunLiveItemListener",
-    "mq_rss_run_parser": "ts-node -r tsconfig-paths/register ./src/index.ts mqRSSRunParser",
-    "mq_rss_run_dlq_consumer": "ts-node -r tsconfig-paths/register ./src/index.ts mqRSSRunDlqConsumer",
-    "orm_feed_update_flag_status": "ts-node -r tsconfig-paths/register ./src/index.ts ormFeedUpdateFlagStatus",
-    "orm_on_demand_parser_event_delete_outdated": "ts-node -r tsconfig-paths/register ./src/index.ts deleteOutdatedOnDemandParserEvent",
-    "orm_on_demand_parser_event_generate_reports": "ts-node -r tsconfig-paths/register ./src/index.ts generateOnDemandParserEventReports",
-    "parser_rss_parse_feed": "ts-node -r tsconfig-paths/register ./src/index.ts parserRSSParseFeed",
-    "podcast_index_dead_feeds_delete_cache": "ts-node -r tsconfig-paths/register ./src/index.ts podcastIndexDeadFeedsDeleteCache",
-    "podcast_index_dead_feeds_flag_and_merge": "ts-node -r tsconfig-paths/register ./src/index.ts podcastIndexDeadFeedsFlagAndMerge",
-    "podcast_index_trending_podcasts_get": "ts-node -r tsconfig-paths/register ./src/index.ts podcastIndexTrendingPodcastsGet",
-    "podcast_index_value_update_all": "ts-node -r tsconfig-paths/register ./src/index.ts podcastIndexValueUpdateAll",
-    "stats_update_aggregated": "ts-node -r tsconfig-paths/register ./src/index.ts statsUpdateAggregated"
+    "archive_all": "node ./dist/index.js archiveAll",
+    "mq_rss_add": "node ./dist/index.js mqRSSAdd",
+    "mq_rss_add_all": "node ./dist/index.js mqRSSAddAll",
+    "mq_rss_add_recently_updated_feeds_from_podcast_index": "node ./dist/index.js mqRSSAddRecentlyUpdatedFeedsFromPodcastIndex",
+    "mq_rss_run_live_item_listener": "node ./dist/index.js mqRSSRunLiveItemListener",
+    "mq_rss_run_parser": "node ./dist/index.js mqRSSRunParser",
+    "mq_rss_run_dlq_consumer": "node ./dist/index.js mqRSSRunDlqConsumer",
+    "orm_feed_update_flag_status": "node ./dist/index.js ormFeedUpdateFlagStatus",
+    "orm_on_demand_parser_event_delete_outdated": "node ./dist/index.js deleteOutdatedOnDemandParserEvent",
+    "orm_on_demand_parser_event_generate_reports": "node ./dist/index.js generateOnDemandParserEventReports",
+    "parser_rss_parse_feed": "node ./dist/index.js parserRSSParseFeed",
+    "podcast_index_dead_feeds_delete_cache": "node ./dist/index.js podcastIndexDeadFeedsDeleteCache",
+    "podcast_index_dead_feeds_flag_and_merge": "node ./dist/index.js podcastIndexDeadFeedsFlagAndMerge",
+    "podcast_index_trending_podcasts_get": "node ./dist/index.js podcastIndexTrendingPodcastsGet",
+    "podcast_index_value_update_all": "node ./dist/index.js podcastIndexValueUpdateAll",
+    "stats_update_aggregated": "node ./dist/index.js statsUpdateAggregated"
   },
   "license": "AGPL-3.0",
   "dependencies": {
-    "@podverse/external-services": "*",
+    "@podverse/external-services-firebase": "*",
+    "@podverse/external-services-podcast-index": "*",
     "@podverse/helpers": "*",
     "@podverse/mq": "*",
     "@podverse/notifications": "*",
