+++ docs/ARCHITECTURE.md
@@ -66,22 +66,22 @@ sequenceDiagram
 
 ## Module Dependency Order
 
-| Tier | Packages                                                                               | Depends On                                          |
-| ---- | -------------------------------------------------------------------------------------- | --------------------------------------------------- |
-| 1    | helpers                                                                                | (none)                                              |
-| 2    | helpers-validation, helpers-requests, helpers-backend, helpers-browser, helpers-config | helpers                                             |
-| 3    | external-services, orm                                                                 | helpers, helpers-\*                                 |
-| 4    | notifications, parser                                                                  | helpers, helpers-\*, external-services, orm         |
-| 5    | mq                                                                                     | helpers, helpers-\*, external-services, orm, parser |
-| 6    | api, web, workers, management-\*                                                       | various                                             |
-| 7    | qa                                                                                     | helpers, helpers-\*, external-services, orm, parser |
+| Tier | Packages                                                                                   | Depends On                                                                            |
+| ---- | ------------------------------------------------------------------------------------------ | ------------------------------------------------------------------------------------- |
+| 1    | helpers                                                                                    | (none)                                                                                |
+| 2    | helpers-validation, helpers-requests, helpers-backend, helpers-browser, helpers-config     | helpers                                                                               |
+| 3    | external-services-firebase, external-services-paypal, external-services-podcast-index, orm | helpers, helpers-\*                                                                   |
+| 4    | notifications, parser                                                                      | helpers, helpers-\*, external-services-firebase, external-services-podcast-index, orm |
+| 5    | mq                                                                                         | helpers, helpers-\*, external-services-podcast-index, orm, parser                     |
+| 6    | api, web, workers, management-\*                                                           | various                                                                               |
+| 7    | qa                                                                                         | helpers, helpers-\*, external-services-\*, orm, parser                                |
 
 ## Build Order
 
 ```
 1. helpers
 2. helpers-validation, helpers-requests, helpers-backend, helpers-browser, helpers-config (parallel)
-3. external-services
+3. external-services-firebase, external-services-paypal, external-services-podcast-index (parallel)
 4. orm
 5. notifications
 6. parser
@@ -102,7 +102,9 @@ packages/                 # Publishable npm packages (@podverse/*)
   helpers-backend/        # Backend-specific utilities (logger, timers, OS)
   helpers-browser/        # Browser-specific utilities (clipboard)
   helpers-config/         # Configuration validation utilities
-  external-services/      # Third-party API integrations
+  external-services-firebase/   # Firebase (Google) integration
+  external-services-paypal/     # PayPal integration
+  external-services-podcast-index/  # Podcast Index API integration
   orm/                    # Database entities, services, migrations
   notifications/          # Push notification services (Firebase)
   parser/                 # RSS/Podcast feed parsing
