+++ package.json
@@ -49,8 +49,8 @@
     "prepare": "bash scripts/git-hooks/install-hooks.sh",
     "start-feature": "bash scripts/start-feature.sh",
     "build": "npm run build --workspaces --if-present",
-    "build:packages": "npm run build -w packages/helpers && npm run build -w packages/helpers-validation && npm run build -w packages/helpers-requests && npm run build -w packages/helpers-backend && npm run build -w packages/helpers-browser && npm run build -w packages/helpers-config && npm run build -w packages/external-services && npm run build -w packages/orm && npm run build -w packages/notifications && npm run build -w packages/parser && npm run build -w packages/mq",
-    "build:packages:prod": "npm run build:prod -w packages/helpers && npm run build:prod -w packages/helpers-validation && npm run build:prod -w packages/helpers-requests && npm run build:prod -w packages/helpers-backend && npm run build:prod -w packages/helpers-browser && npm run build:prod -w packages/helpers-config && npm run build:prod -w packages/external-services && npm run build:prod -w packages/orm && npm run build:prod -w packages/notifications && npm run build:prod -w packages/parser && npm run build:prod -w packages/mq",
+    "build:packages": "npm run build -w packages/helpers && npm run build -w packages/helpers-validation && npm run build -w packages/helpers-requests && npm run build -w packages/helpers-backend && npm run build -w packages/helpers-browser && npm run build -w packages/helpers-config && npm run build -w packages/external-services-firebase && npm run build -w packages/external-services-paypal && npm run build -w packages/external-services-podcast-index && npm run build -w packages/orm && npm run build -w packages/notifications && npm run build -w packages/parser && npm run build -w packages/mq",
+    "build:packages:prod": "npm run build:prod -w packages/helpers && npm run build:prod -w packages/helpers-validation && npm run build:prod -w packages/helpers-requests && npm run build:prod -w packages/helpers-backend && npm run build:prod -w packages/helpers-browser && npm run build:prod -w packages/helpers-config && npm run build:prod -w packages/external-services-firebase && npm run build:prod -w packages/external-services-paypal && npm run build:prod -w packages/external-services-podcast-index && npm run build:prod -w packages/orm && npm run build:prod -w packages/notifications && npm run build:prod -w packages/parser && npm run build:prod -w packages/mq",
     "build:apps": "npm run build --workspaces --if-present --workspace=apps/api --workspace=apps/web --workspace=apps/workers --workspace=apps/management-api --workspace=apps/management-web",
     "clean": "npm run clean --workspaces --if-present",
     "clean:packages": "find packages -name 'tsconfig.tsbuildinfo' -delete && find packages -type d -name 'dist' -exec rm -rf {} + 2>/dev/null || true",
@@ -60,7 +60,7 @@
     "clean:node_modules": "rm -rf node_modules packages/*/node_modules apps/*/node_modules tools/*/node_modules",
     "reinstall": "npm run clean:node_modules && rm -f package-lock.json && npm install",
     "reinstall:full": "npm run clean:all && npm run reinstall && npm run build:packages && npm run build:apps",
-    "watch:packages": "concurrently -n helpers,h-valid,h-req,h-back,h-brow,h-conf,ext-svc,orm,notif,parser,mq -c blue,yellow,magenta,cyan,green,red,blueBright,greenBright,magentaBright,cyanBright,yellowBright \"npm run build:watch -w packages/helpers\" \"sleep 1 && npm run build:watch -w packages/helpers-validation\" \"sleep 2 && npm run build:watch -w packages/helpers-requests\" \"sleep 3 && npm run build:watch -w packages/helpers-backend\" \"sleep 4 && npm run build:watch -w packages/helpers-browser\" \"sleep 5 && npm run build:watch -w packages/helpers-config\" \"sleep 6 && npm run build:watch -w packages/external-services\" \"sleep 9 && npm run build:watch -w packages/orm\" \"sleep 12 && npm run build:watch -w packages/notifications\" \"sleep 15 && npm run build:watch -w packages/parser\" \"sleep 18 && npm run build:watch -w packages/mq\"",
+    "watch:packages": "concurrently -n helpers,h-valid,h-req,h-back,h-brow,h-conf,ext-fb,ext-pay,ext-pi,orm,notif,parser,mq -c blue,yellow,magenta,cyan,green,red,blueBright,greenBright,magentaBright,cyanBright,yellowBright,gray,red \"npm run build:watch -w packages/helpers\" \"sleep 1 && npm run build:watch -w packages/helpers-validation\" \"sleep 2 && npm run build:watch -w packages/helpers-requests\" \"sleep 3 && npm run build:watch -w packages/helpers-backend\" \"sleep 4 && npm run build:watch -w packages/helpers-browser\" \"sleep 5 && npm run build:watch -w packages/helpers-config\" \"sleep 6 && npm run build:watch -w packages/external-services-firebase\" \"sleep 7 && npm run build:watch -w packages/external-services-paypal\" \"sleep 8 && npm run build:watch -w packages/external-services-podcast-index\" \"sleep 9 && npm run build:watch -w packages/orm\" \"sleep 12 && npm run build:watch -w packages/notifications\" \"sleep 15 && npm run build:watch -w packages/parser\" \"sleep 18 && npm run build:watch -w packages/mq\"",
     "dev:api": "npm run dev -w apps/api",
     "dev:api:watch": "npm run dev:watch -w apps/api",
     "dev:web": "npm run dev -w apps/web",
