+++ packages/external-services-firebase/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@podverse/notifications",
+  "name": "@podverse/external-services-firebase",
   "version": "5.2.5",
-  "description": "Push notification helper module for Podverse use cases",
+  "description": "Firebase (Google) integration for Podverse",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "files": [
@@ -21,14 +21,8 @@
     "access": "public"
   },
   "dependencies": {
-    "@podverse/external-services-firebase": "*",
     "@podverse/helpers": "*",
-    "web-push": "^3.6.7",
-    "@podverse/helpers-backend": "*",
-    "@podverse/helpers-validation": "*"
-  },
-  "devDependencies": {
-    "@types/web-push": "^3.6.4"
+    "firebase-admin": "^13.6.0"
   },
   "type": "module"
 }
