+++ packages/external-services-firebase/src/factory.ts
