+++ packages/external-services-firebase/src/services/google/firebase/index.ts
