+++ packages/external-services-firebase/tsconfig.json
@@ -7,11 +7,9 @@
     "moduleResolution": "NodeNext",
     "outDir": "./dist",
     "rootDir": "./src",
-    "experimentalDecorators": true,
-    "emitDecoratorMetadata": true,
     "baseUrl": "./"
   },
-  "include": ["./src/**/*.ts"],
+  "include": ["./src/**/*.ts", "./src/**/*.d.ts"],
   "exclude": ["node_modules", "dist"],
-  "references": [{ "path": "../helpers" }, { "path": "../external-services-firebase" }]
+  "references": [{ "path": "../helpers" }]
 }
