+++ packages/external-services-podcast-index/package.json
@@ -1,15 +1,15 @@
 {
-  "name": "@podverse/external-services",
+  "name": "@podverse/external-services-podcast-index",
   "version": "5.2.5",
-  "description": "External service integrations for Podverse",
+  "description": "Podcast Index API integration for Podverse",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "files": [
     "dist/**/*"
   ],
   "scripts": {
-    "build": "npm run lint && tsc && tsc-alias",
-    "build:prod": "tsc && tsc-alias",
+    "build": "npm run lint && tsc",
+    "build:prod": "tsc",
     "build:watch": "tsc --build --watch --preserveWatchOutput",
     "lint": "eslint ./src --max-warnings 0",
     "lint:fix": "eslint ./src --fix",
@@ -24,16 +24,7 @@
     "@podverse/helpers": "*",
     "@podverse/helpers-backend": "*",
     "@podverse/helpers-requests": "*",
-    "csv-parser": "^3.2.0",
-    "firebase-admin": "^13.6.0",
-    "http-errors": "2.0.1",
-    "paypal-rest-sdk": "2.0.0-rc.2",
-    "web-push": "^3.6.7"
-  },
-  "devDependencies": {
-    "@types/http-errors": "^2.0.5",
-    "@types/web-push": "^3.6.4",
-    "tsc-alias": "^1.8.16"
+    "csv-parser": "^3.2.0"
   },
   "type": "module"
 }
