+++ packages/external-services-podcast-index/tsconfig.json
@@ -7,11 +7,8 @@
     "moduleResolution": "NodeNext",
     "outDir": "./dist",
     "rootDir": "./src",
-    "experimentalDecorators": true,
-    "emitDecoratorMetadata": true,
     "baseUrl": "./"
   },
-  "include": ["./src/**/*.ts"],
-  "exclude": ["node_modules", "dist"],
-  "references": [{ "path": "../helpers" }, { "path": "../external-services-firebase" }]
+  "include": ["./src/**/*.ts", "./src/**/*.d.ts"],
+  "exclude": ["node_modules", "dist"]
 }
