--- packages/external-services/src/@types/modules.d.ts
@@ -1,4 +1,0 @@
-declare module "paypal-rest-sdk" {
-  const paypalRestSdk: any;
-  export default paypalRestSdk;
-}
