--- packages/external-services/src/config/index.ts
@@ -1,2 +1,0 @@
-// Re-export config types for app-level use
-export * from './types.js';
