--- packages/external-services/src/module-alias-config.ts
@@ -1,12 +1,0 @@
-import moduleAlias from 'module-alias';
-import path from 'path';
-import { fileURLToPath } from 'url';
-
-const __filename = fileURLToPath(import.meta.url);
-const __dirname = path.dirname(__filename);
-
-moduleAlias.addAliases({
-  '@external-services': path.join(__dirname, ''),
-});
-
-export {};
