--- packages/external-services/src/services/index.ts
@@ -1,3 +1,0 @@
-export * from './google/firebase/index.js';
-export * from './paypal/index.js';
-export * from './podcast-index/index.js';
