+++ packages/mq/package.json
@@ -21,7 +21,7 @@
     "access": "public"
   },
   "dependencies": {
-    "@podverse/external-services": "*",
+    "@podverse/external-services-podcast-index": "*",
     "@podverse/helpers": "*",
     "@podverse/helpers-backend": "*",
     "@podverse/helpers-requests": "*",
