+++ packages/mq/src/functions/mq/rss/addRecentlyUpdatedFeedsFromPodcastIndex.ts
@@ -1,4 +1,4 @@
-import type { PodcastIndexService } from '@podverse/external-services';
+import type { PodcastIndexService } from '@podverse/external-services-podcast-index';
 import type { MQQueueConfigFunctionParams } from '@podverse/helpers';
 import { FeedService } from '@podverse/orm';
 import type { ActiveMQArtemisService } from '@queue/services/activeMQArtemis/index.js';
