+++ packages/mq/tsconfig.json
@@ -18,7 +18,7 @@
   "exclude": ["node_modules", "dist"],
   "references": [
     { "path": "../helpers" },
-    { "path": "../external-services" },
+    { "path": "../external-services-podcast-index" },
     { "path": "../orm" },
     { "path": "../parser" }
   ]
