+++ packages/parser/package.json
@@ -21,7 +21,8 @@
     "access": "public"
   },
   "dependencies": {
-    "@podverse/external-services": "*",
+    "@podverse/external-services-firebase": "*",
+    "@podverse/external-services-podcast-index": "*",
     "@podverse/helpers": "*",
     "@podverse/helpers-backend": "*",
     "@podverse/helpers-requests": "*",
