+++ packages/parser/src/context.ts
@@ -1,5 +1,6 @@
 import type { ILoggerLike, TimerManager } from '@podverse/helpers-backend';
-import type { PodcastIndexService, FirebaseContext } from '@podverse/external-services';
+import type { FirebaseContext } from '@podverse/external-services-firebase';
+import type { PodcastIndexService } from '@podverse/external-services-podcast-index';
 import type { NotificationsContext } from '@podverse/notifications';
 import type { ParserConfig } from './config/types.js';
